backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
kombu==5.3.7
packaging==24.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
waitress==3.0.0
wcwidth==0.2.13
wheel==0.43.0
redis
gunicorn